requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
bleach==2.1.2
certifi==2017.11.5
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
configparser==3.5.0
constantly==15.1.0
cryptography==2.1.4
cssselect==1.0.3
decorator==4.1.2
Django==1.8
django-cors-headers==2.4.0
djangorestframework==3.3.0
entrypoints==0.2.3
enum34==1.1.6
functools32==3.2.3.post2
html5lib==1.0.1
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
ipaddress==1.0.19
ipykernel==4.7.0
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==7.1.0
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.1
jupyter-console==5.2.0
jupyter-core==4.4.0
lxml==4.1.1
MarkupSafe==1.0
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.2.2
pandocfilters==1.4.2
parsel==1.4.0
pathlib2==2.3.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
Pygments==2.2.0
pyOpenSSL==17.5.0
python-dateutil==2.6.1
pytz==2018.4
pyzmq==16.0.3
qtconsole==4.3.1
queuelib==1.4.2
request==0.0.22
requests==2.18.4
scandir==1.6
Scrapy==1.5.0
service-identity==17.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
testpath==0.3.1
tornado==4.5.3
traitlets==4.3.2
Twisted==17.9.0
urllib3==1.22
w3lib==1.19.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.1.0
win-unicode-console==0.5
wincertstore==0.2
zope.interface==4.4.3
